Remove user interupt from the HomeWIndow, and display activation screen.
The method wait_user_resume(), from the interface UIHandlerOfConnection as been
removed: It was difficult to catch all errors in the connection proccess, and
user interuptions are completely hidden by the future not returning. In other
words, it's an implementation details of the GUI.

The HomeWindow has now a mthod to changes the current screen, nd switch between
screens.

The Activation screen send an events when the user use one of the two buttons,
and the home window acts in consequence.