Move all code related to passphrase from BajooApp to a new class PassphraseManager.
PassphraseManager is in charge of asking the user's passphrase, storing it and
giving it to the encryption module. As usual, BajooApp makes the connection
between the encryption module, the passphrase manager and the passphrase window.

Also, the PassphraseManager detect calls done after the user has refused to
gives his passphrase, and ignore them.