Adapt filesync to use the task consumer; Remove Future class.
The module filesync has been rewritten to use the task_consumer module.
The LocalContainer acquire_index() method now use a promise instead of
callbacks, which is more coherent with the rest of the code.
However, these two part (filesync module and index managment) are still way
much complicated, and should be rewritten.

Also, as filesync was the last module who depends on Future, the Future module
is now removed.