Add SoftwareUpdater class.
This class is aimed to replace the poorly-integrated wx.SoftwareUpdate.

The new update class don't display window; All operations are done in
background. The app is restarted as soon as the update is done.

Update checks are done daily, even if the app is not restarted.

Note that the class doesn't fully replace (yet) wx.SoftwareUpdate. The
integration with the GUI will be implemented latter.