Rewrite TaskBarIcon module to a more lisible, MVC-based architecture.
All public methods are exposed by the Controller class.

The task bar icon should be instanciated from the `make_task_bar_icon` method,
which will use the appriopriate view.

`base.py` describe the interface between controller and views.
`controller.py` is the only implementation of the controler, and can use two
views: the wxPython one and the Unity one.

Both views shares a lot of data, located in `common_view_data.py`. It includes
paths of icon files and an abstact representation of menus.

The Unity view uses a distinct process. The controller don't communicate
directly to the real unity view, but use an adapter instead.