Add an option to detect uncaught errors on delayed GUI actions.
`@ensure_gui_thread` is used to call a GUI action from another thread.
The GUI action is wrapped in a Promise object.
There is a lot of calls to @ensure_gui_thread done without checking the
result. GUI errors was silently ignored.

This commit add a 'safeguard' option to display ignored errors. It follow
the same convention as '@reduce_coroutine'.