[New Index Tree] Add a new IndexTree structure to deal with the BaseNode.
Add the task attribute in BaseNode: a task is any object (or reference)
that is in charge of the sync of the node. A node can have only one task
at the same time.

The IndexTree can browse through the nodes, and find all nodes that need
to be sync. It handles itself the complexity of concurrent accesses.