[New Index Tree] Add SyncScheduler class, to browse many index trees at once.
The SyncScheduler browses over a set of index tree. It will be used by
ContainerSyncPool to get the next node that should be sync.

Note:
I'm not fully satisfied by this implementation. I think the code is neither
easily readable nor efficient, but I've not found a better way to achieve what I want (browsing tree per tree).
It do the job and there are unit tests, so I'll keep it like this.