Send error notification from ContainerSyncPool instead of tasks.
The code for sending notification when an error occurs in a sync task is
moved from AbstractTask to ContainerSyncPool.
ContainerSyncPool already had some error control methods, and it makes sense
to gather all code related to errors in the same place.

The code logic (reaction to errors) remains unchanged.

Also, note that startup sync tasks will not return errors before the full sync
of the container. It should be fixed in the next commits.