Add a helper method in HintBuilder to break coupled hints.
The move hints explicitly "links" two nodes together:
Acquire a node which is either the source or the destination of a move
also implies acquiring the second node.

Some task can't handle both nodes, and the node's link(s) must be broken
before the task can acquire the node. This method provides an easy way
to do this, while keeping a coherent state after the link's break.