HintBuilder: make methods taking node in parameter instead of path.
The methods "by path" handles the IndexTree, and are appropriate for single
atomic actions.
The new methods "by node" are better suited for an use as a part of a group of
action over the tree/nodes.

Also, an empty node with a task assigned is no more removed from the tree,
in case of "delete" hint.