Add "unlock()" method to LocalContainer
This method will ensures the container key (and so the container as a whole) is
usable by the sync pool. The method is used before any addition of a container
in the sync pool.
As a result, the "Container.error" attribute has no reason to exists anymore,
and the error-handling code of ContainerSyncPool is simplified.