Retry node sync after an error
Nodes in error are stored in a special list and "removed" from the normal sync
path. A delay before the next sync attempt is defined (depending of the number
of try). The node will be sync again (in priority) after the delay elapses.

An IndexTree has the status "sync" when all its nodes are synced **except nodes
in error**. The "sync" status indicates that there is no more activity on this
tree, but the tree can be not fully sync.